Fix write command
