+++ python/can_bridge/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['frame']
+__all__ = ['frame', 'commands']
